builder: save function.
Note that this new save function does not create a new layer, this is
(forgotten by me) different with the tag verb which does. I may leave
the tag verb alone now because of this and add the save verb which
operates slightly different and will combine parameters to form one
function. This is instrinsically different than tag which just saves a
layer to docker.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>