builder,fetcher,layers: fix cmd/entrypoint semantics
previously, nulls would be set on the entrypoint, triggering inheritance
from `run` statments that were the previous commit in the docker chain.
What this does now is set it to `[]string` which is an empty command in
the docker configuration.

Note that there is now no way to do `entrypoint nil` in a way that will
inherit from the previous entrypoint statement, however, I don't think
this is necessarily a bad thing. :)

Signed-off-by: Erik Hollensbe <github@hollensbe.org>