portable.sh.tmpl: bring in environment variables from the host
Previously, if the user specified environment variables, they were
ignored by the portable script. This is because the environment
variables would need to have been propagated to docker that box runs in,
so it can then read the environment variables from the container.

Now, you can specify BOX_INCLUDE_ENV to whitespace-delimit environment
variables to pass from your host into the container, allowing it to be
easy to generically do this in make tasks etc.

This is a stop-gap for environment usage on OS X in particular, until
0.6.0 arrives (with real dictionaries) this feature is really
needed.

This will be backported into 0.5.4.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>